+++ src/gstore/gmessagestore.cpp
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2001-2024 Graeme Walker <graeme_walker@users.sourceforge.net>
+// Copyright (C) 2026 Graeme Walker <graeme_walker@users.sourceforge.net>
 // 
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -32,7 +32,10 @@ std::unique_ptr<GStore::StoredMessage> GStore::operator++( std::shared_ptr<Messa
 
 GStore::MessageStore::AddressStyle GStore::MessageStore::addressStyle( std::string_view address )
 {
-	if( address.empty() || address.at(0U) == '@' || address.at(address.size()-1U) == '@' )
+	if( address.empty() )
+		return AddressStyle::Empty ;
+
+	if( address.at(0U) == '@' || address.at(address.size()-1U) == '@' )
 		return AddressStyle::Invalid ; // missing stuff
 
 	if( !G::Str::isPrintable(address) )